+++ tests/internal/invalid/upgradeable/use_comprehension/1.py
@@ -1,0 +1,3 @@
+operands = list(b8(token) for token in tokens)
+patterns = dict((token.name, b8(token)) for token in tokens)
+unique_operands = set(b8(token) for token in tokens)