+++ tests/internal/valid/upgradeble/list_comp/1.py
@@ -1,3 +1,0 @@
-def test_reversed(self):
-    input = list(map(ord, "Hello"))
-    input = list(map(my.custom.ord, "Hello"))