+++ tests/internal/valid/upgradeble/use_comprehension/1.py
@@ -1,0 +1,1 @@
+dict(token.expand() for token in tokens)